backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiosqlite==0.20.0
python-dotenv==1.0.1
sentence-transformers==3.3.1
numpy>=1.24.0
httpx>=0.27.0
